+++ functions/transitions.function.ts
@@ -1,0 +1,8 @@
+let transitionsMap = {};
+
+export function transitions() {
+    return {
+        get: (transition) => transitionsMap[transition],
+        set: (name, transition) => { transitionsMap = Object.assign({}, transitionsMap, { [name]: transition }); }
+    }
+}