+++ polyfills/load-polyfills.function.ts
@@ -1,0 +1,5 @@
+import { objectAssignPoly } from "./object-assign.poly";
+
+export function loadPolyfills() {
+    objectAssignPoly();
+}